+++ server/src/main/scala/org/apache/livy/server/LivyServer.scala
@@ -222,7 +222,7 @@ class LivyServer extends Logging {
               mount(context, staticResourceServlet, "/static/*")
               mount(context, uiRedirectServlet(basePath + "/ui/"), "/*")
               thriftServerFactory.foreach { factory =>
-                mount(context, factory.getUI(basePath), factory.getUIMappings: _*)
+                mount(context, factory.getServlet(basePath), factory.getServletMappings: _*)
               }
             } else {
               mount(context, uiRedirectServlet(basePath + "/metrics"), "/*")
