+++ thriftserver/server/src/main/scala/org/apache/livy/thriftserver/ui/ThriftJsonServlet.scala
@@ -19,46 +19,20 @@ package org.apache.livy.thriftserver.ui
 
 import java.text.SimpleDateFormat
 
-import scala.xml.Node
-
 import org.apache.livy.server.JsonServlet
-import org.apache.livy.server.ui.BaseServlet
 import org.apache.livy.thriftserver.LivyThriftServer
 
 
-class ThriftUIServlet(val basePath: String) extends JsonServlet with BaseServlet {
+class ThriftJsonServlet(val basePath: String) extends JsonServlet {
 
   private val df = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss")
-
-  private case class ThriftSessionsPage(name: String = "JDBC/ODBC Sessions") extends Page {
-    override def getNavTabs: Seq[Node] = {
-      <li><a href={basePath + "/ui"}>Sessions</a></li> ++
-        <li class="active"><a href="#">Thrift Server</a></li>
-    }
-  }
 
   case class SessionInfo(sessionId: String,
     livySessionId: String,
     owner: String,
     createdAt: String)
 
-  def withContentType(ct: String)(f: Any): Any = {
-    val fullContentType = formats(ct)
-    contentType = fullContentType
-    response.setContentType(fullContentType)
-    f
-  }
-
-  get("/") { withContentType("html") {
-    val content =
-      <div id="thrift-sessions">
-        <script src={s"$basePath/static/js/thrift-sessions.js"}></script>
-      </div>
-
-    createPage(ThriftSessionsPage(), content)
-  }}
-
-  get("/sessions") { withContentType("json") {
+  get("/sessions") {
     val thriftSessions = LivyThriftServer.getInstance.map { server =>
       val sessionManager = server.getSessionManager()
       sessionManager.getSessions.map { sessionHandle =>
@@ -76,5 +50,5 @@ class ThriftUIServlet(val basePath: String) extends JsonServlet with BaseServlet
       "from" -> from,
       "total" -> thriftSessions.length,
       "sessions" -> thriftSessions.view(from, from + size))
-  }}
+  }
 }
